try to fix SQL query that causes product image not showing (#2)

+++ pages/product/SimplePager.php
@@ -12,14 +12,14 @@ class SimplePager {
         global $_db;
         
         // Set [limit] and [page]
-        $this->limit = ctype_digit($limit) ? max($limit, 1) : 10;
-        $this->page = ctype_digit($page) ? max($page, 1) : 1;
+        $this->limit = ctype_digit($limit) ? (int)max($limit, 1) : 10;
+        $this->page = ctype_digit($page) ? (int)max($page, 1) : 1;
 
         // Set [item count]
         $q = preg_replace('/SELECT.+FROM/', 'SELECT COUNT(*) FROM', $query, 1);
         $stm = $_db->prepare($q);
         $stm->execute($params);
-        $this->item_count = $stm->fetchColumn();
+        $this->item_count = (int)$stm->fetchColumn();
 
         // Set [page count]
         $this->page_count = ceil($this->item_count / $this->limit);
